renames protocol into p2p
