+++ network/p2p/conduit.go
@@ -1,4 +1,4 @@
-package protocol
+package p2p
 
 import (
 	"context"
