+++ network/p2p/connGator.go
@@ -1,4 +1,4 @@
-package protocol
+package p2p
 
 import (
 	"sync"
