+++ network/p2p/connManager.go
@@ -1,4 +1,4 @@
-package protocol
+package p2p
 
 import (
 	"github.com/libp2p/go-libp2p-core/connmgr"
