+++ network/p2p/libp2pNode_test.go
@@ -1,4 +1,4 @@
-package protocol
+package p2p
 
 import (
 	"bufio"
