+++ network/p2p/libp2pUtils.go
@@ -1,4 +1,4 @@
-package protocol
+package p2p
 
 // All utilities for libp2p not natively provided by the library.
