+++ network/p2p/nodeIDRefresher.go
@@ -1,4 +1,4 @@
-package protocol
+package p2p
 
 import (
 	"github.com/rs/zerolog"
